refactor : 명칭변경
