+++ app/store/useStore.ts
@@ -8,7 +8,7 @@ interface State {
   switch?: boolean;
   spinStop?: boolean;
   recode?: boolean;
-  sound?: boolean;
+  sound?: number;
   play?: boolean;
   lp?: object;
   lpSwitch? : boolean;
@@ -41,7 +41,7 @@ const initialState: LoadingState = {
   main: {
     lp: lpInitState[0],
     lpSwitch : false,
-    sound: true,
+    sound: 30,
     play: false,
   },
 };
@@ -106,9 +106,9 @@ const actions = (set: any, get: any): LoadingAction => ({
   }
 });
 
-const useMotionStore = create<Store>((set, get) => ({
+const useStore = create<Store>((set, get) => ({
   ...initialState,
   ...actions(set, get),
 }));
 
-export default useMotionStore;
+export default useStore;
